refactor: move script files into folders
