+++ src/marshalling/boolean-serializer.ts
@@ -1,7 +1,7 @@
 // in the packed version it can be a bitvector of booleans
 // until then trololo
 
-import { Bytes } from './utils'
+import { Bytes } from '../utils'
 
 export function serializeBoolean(value: boolean): Bytes<1> {
   return new Bytes([Number(value)])
