+++ src/marshalling/string-seralizer.ts
